Disable case-insensitive matches for font-lock mode
We have a problem here php-mode gives a variable such as ‘$_server’
the same treatment it gives ‘$_SERVER’.  These superglobals are
case-sensitive in the eyes of PHP and are not interchangeable.  So
php-mode should not treat them equally either.  The first step to move
in that direction is to disable case-insensitive matching for
font-lock within all aspects of php-mode, enabling us to pick and
choose which parts should continue to use the old case-sensitive
behavior.

Github-issue: 75