Remove magic constants from the ‘php-constants’ variable
Most of the time we want to perform case-sensitive searches for PHP
constants.  The exceptions are the magic constants where we do want
the previous case-insensitive behavior.  To make it easier to treat
those magic constants differently we extract them from ‘php-constants’
and place them in their own variable.  This patch also adds that new
list to ‘php-font-lock-keywords-1’ so that php-mode will still give
them highlighting but it still fails to properly highlight the
lowercase variations of the magic constants.

Github-issue: 75